Improve the procedures experiment and polish up chlib
Improve documentation and organization.  With this experiment, a set
of steps are requested, one for each study in the database.  The
expectation is that another process will be triggered that will
work on performing the requested steps and will insert activity
documents into the database.  The agent is watching a chances feed
that will print these as they are inserted.